fix(ci): python-coverage-shards maintainability — split merge-method test suite under 1500-line limit
tests/unit/runtime/test_pr_monitor_merge_methods.py exceeded the 1500-line
first-party file guardrail (1598 lines). Extract its shared doubles
(constants, _mergeable_status, _MergeMethodClient, _execute_merge) into a new
_merge_methods_fixtures helper module and split the cases into two focused
suites — merge-method selection/preflight (test_pr_monitor_merge_methods)
and merge-failure handling (test_pr_monitor_merge_failures). All 33 cases
preserved; behavior unchanged.
