+++ tests/unit/runtime/test_pr_monitor_merge_methods.py
@@ -1,35 +1,36 @@
-"""Regression tests for PR monitor merge-method selection."""
+"""Regression tests for PR monitor merge-method selection and preflight."""
 
 from __future__ import annotations
 
 from collections.abc import AsyncIterator
-from datetime import UTC, datetime
 from pathlib import Path
 
 import pytest
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
 from awf.common.bitbucket_client import (
-    BITBUCKET_MERGE_IN_PROGRESS,
-    BITBUCKET_MERGE_METHOD_UNSUPPORTED,
-    BITBUCKET_MERGE_TASK_TIMEOUT,
     BITBUCKET_RATE_LIMITED,
     BitbucketClientError,
 )
 from awf.common.commands import FakeCommandRunner
-from awf.common.github_client import GitHubClientError, RepoRef
-from awf.db.enums import OperationStatus, OperationType
+from awf.common.github_client import GitHubClientError
+from awf.db.enums import OperationStatus
 from awf.db.repositories import OperationRepository, WorkspaceEventRepository, WorkspaceRepository
 from awf.db.session import make_session_factory
 from awf.runtime.pr_monitor import (
-    CheckState,
     Merge,
-    MergeableState,
-    MergeStateStatus,
     MonitorState,
-    PRStatus,
 )
 from tests.postgres import postgres_test_engine
+from tests.unit.runtime._merge_methods_fixtures import (
+    _TEST_DEFAULT_BASE_BRANCH,
+    _TEST_MERGE_ONLY_BASE_BRANCH,
+    _TEST_PR_NUMBER,
+    _TEST_REPO,
+    _execute_merge,
+    _mergeable_status,
+    _MergeMethodClient,
+)
 from tests.unit.runtime._monitor_runner_fixtures import (
     FakeAdapter,
     RecordedSleep,
@@ -37,161 +38,12 @@ from tests.unit.runtime._monitor_runner_fixtures import (
     seed_monitoring_workspace,
 )
 
-_TEST_REPO = RepoRef(owner="example-org", name="example-repo")
-_TEST_PR_NUMBER = 42
-_TEST_DEFAULT_BASE_BRANCH = "release/default"
-_TEST_MERGE_ONLY_BASE_BRANCH = "release/merge-only"
-
 
 @pytest.fixture
 async def factory() -> AsyncIterator[async_sessionmaker[AsyncSession]]:
     """Provide an isolated async session factory for merge-method monitor tests."""
     async with postgres_test_engine() as engine:
         yield make_session_factory(engine)
-
-
-def _mergeable_status() -> PRStatus:
-    """Build a clean PR status suitable for merge-loop regression tests."""
-    return PRStatus(
-        number=_TEST_PR_NUMBER,
-        head_sha="abc1234567890def",
-        mergeable=MergeableState.MERGEABLE,
-        check_state=CheckState.SUCCESS,
-        unresolved_inline_threads=(),
-        unresolved_review_comments=(),
-        blocking_reviews=(),
-        base_behind_count=0,
-        merge_state_status=MergeStateStatus.CLEAN,
-        checks=(),
-    )
-
-
-class _MergeMethodClient:
-    """GitHub client test double for merge-method policy scenarios."""
-
-    def __init__(
-        self,
-        *,
-        repo_methods: tuple[str, ...] = ("merge", "squash", "rebase"),
-        branch_methods: tuple[str, ...] | None = None,
-        repo_error: GitHubClientError | None = None,
-        branch_error: GitHubClientError | None = None,
-        merge_results: list[str | GitHubClientError | BitbucketClientError] | None = None,
-        post_comment_error: GitHubClientError | BitbucketClientError | None = None,
-    ) -> None:
-        """Configure repository policy, branch policy, and merge outcomes."""
-        self.repo_methods = repo_methods
-        self.branch_methods = branch_methods
-        self.repo_error = repo_error
-        self.branch_error = branch_error
-        self.merge_results = merge_results or ["MERGESHA123"]
-        self.post_comment_error = post_comment_error
-        self.merge_calls: list[str] = []
-        self.comments: list[str] = []
-        self.expected_repo = _TEST_REPO
-        self.expected_pr_number = _TEST_PR_NUMBER
-        self.expected_base_branch = _TEST_DEFAULT_BASE_BRANCH
-
-    def expect_context(
-        self,
-        *,
-        repo: RepoRef,
-        pr_number: int,
-        base_branch: str,
-    ) -> None:
-        """Set expected repository, PR, and base branch for assertions."""
-        self.expected_repo = repo
-        self.expected_pr_number = pr_number
-        self.expected_base_branch = base_branch
-
-    async def fetch_repo_merge_methods(self, *, repo: RepoRef) -> tuple[str, ...]:
-        """Return configured repository merge methods or raise the configured error."""
-        assert repo == self.expected_repo
-        if self.repo_error is not None:
-            raise self.repo_error
-        return self.repo_methods
-
-    async def fetch_branch_pull_request_allowed_merge_methods(
-        self,
-        *,
-        repo: RepoRef,
-        branch: str,
-    ) -> tuple[str, ...] | None:
-        """Return configured branch merge methods or raise the configured error."""
-        assert repo == self.expected_repo
-        assert branch == self.expected_base_branch
-        if self.branch_error is not None:
-            raise self.branch_error
-        return self.branch_methods
-
-    async def merge_pr(
-        self,
-        *,
-        repo: RepoRef,
-        pr_number: int,
-        method: str = "squash",
-        delete_branch: bool = True,
-    ) -> str:
-        """Record the attempted method and return or raise the queued outcome."""
-        assert repo == self.expected_repo
-        assert pr_number == self.expected_pr_number
-        assert delete_branch is True
-        self.merge_calls.append(method)
-        result = self.merge_results.pop(0)
-        if isinstance(result, GitHubClientError | BitbucketClientError):
-            raise result
-        return result
-
-    async def post_comment(self, *, repo: RepoRef, pr_number: int, body: str) -> None:
-        """Record human notification comments emitted by the merge loop."""
-        assert repo == self.expected_repo
-        assert pr_number == self.expected_pr_number
-        if self.post_comment_error is not None:
-            raise self.post_comment_error
-        self.comments.append(body)
-
-
-async def _execute_merge(
-    *,
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-    gh: _MergeMethodClient,
-    base_branch: str = _TEST_DEFAULT_BASE_BRANCH,
-) -> tuple[bool | None, MonitorState, RecordedSleep, str]:
-    """Seed a monitored workspace and execute one merge action."""
-    workspace_id = await seed_monitoring_workspace(factory)
-    sleep_fn = RecordedSleep()
-    gh.expect_context(
-        repo=_TEST_REPO,
-        pr_number=_TEST_PR_NUMBER,
-        base_branch=base_branch,
-    )
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=FakeAdapter(),
-        sleep_fn=sleep_fn,
-        worktrees_root=tmp_path / "worktrees",
-        gh=gh,
-    )
-    state = MonitorState()
-
-    terminal = await runner._execute(
-        action=Merge(),
-        workspace_id=workspace_id,
-        repo_url=f"git@github.com:{_TEST_REPO.slug()}.git",
-        repo=_TEST_REPO,
-        pr_number=_TEST_PR_NUMBER,
-        status=_mergeable_status(),
-        state=state,
-        base_branch=base_branch,
-        remote_branch=f"awf/{workspace_id}",
-        remote_push_url=f"git@github.com:{_TEST_REPO.slug()}.git",
-        compose_project="proj",
-        compose_file=tmp_path / "compose.yml",
-        monitor_log=None,
-    )
-    return terminal, state, sleep_fn, workspace_id
 
 
 @pytest.mark.unit
@@ -829,770 +681,3 @@ async def test_non_rebase_merge_with_empty_merge_commit_records_head_marker(
     assert not any(
         key.startswith("__awf_merge_method_blocked__:") for key in state.threads_addressed_ids
     )
-
-
-@pytest.mark.unit
-async def test_transient_first_merge_failure_does_not_retry_allowed_alternative(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """A transient merge failure backs off instead of changing merge method."""
-    gh = _MergeMethodClient(
-        repo_methods=("merge", "squash"),
-        branch_methods=("merge", "squash"),
-        merge_results=[
-            GitHubClientError(
-                operation="gh pr merge",
-                returncode=1,
-                stderr="HTTP 502 Bad Gateway",
-            ),
-            "MERGESHA789",
-        ],
-    )
-
-    terminal, state, sleep_fn, _workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    assert gh.merge_calls == ["squash"]
-    assert sleep_fn.calls == [5]
-    assert gh.comments == []
-    assert not any(
-        key.startswith("__awf_merge_method_blocked__:") for key in state.threads_addressed_ids
-    )
-
-
-@pytest.mark.unit
-async def test_exhausted_transient_merge_failure_preserves_retry_counter(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """An exhausted *transient* merge blip must not reset the ``merge_pr`` budget.
-
-    When the bounded forge-retry helper returns ``False`` because the budget is
-    exhausted (the merge error is still transient, not a deterministic rejection),
-    the merge arm falls back to notify-and-keep-polling. It must keep the persisted
-    ``merge_pr`` retry counter so later polls fail closed via the exhausted path,
-    rather than clearing it as if the blocker were deterministic and handing each
-    subsequent merge attempt a fresh full retry budget — symmetric with
-    ``fetch_pr_status`` / ``pre_merge_recheck`` (regression for the PR #516
-    merge-path counter-reset bug).
-    """
-    workspace_id = await seed_monitoring_workspace(factory)
-    sleep_fn = RecordedSleep()
-    gh = _MergeMethodClient(
-        repo_methods=("squash",),
-        branch_methods=("squash",),
-        merge_results=[
-            GitHubClientError(
-                operation="gh pr merge",
-                returncode=1,
-                stderr="HTTP 502 Bad Gateway",
-            )
-        ],
-    )
-    gh.expect_context(
-        repo=_TEST_REPO,
-        pr_number=_TEST_PR_NUMBER,
-        base_branch=_TEST_DEFAULT_BASE_BRANCH,
-    )
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=FakeAdapter(),
-        sleep_fn=sleep_fn,
-        worktrees_root=tmp_path / "worktrees",
-        gh=gh,
-    )
-    # Zero retries: the first transient blip exhausts the budget immediately and
-    # drives the merge arm down the notify-and-keep-polling fallback.
-    object.__setattr__(runner._runner_config, "transient_forge_max_retries", 0)
-    state = MonitorState()
-
-    terminal = await runner._execute(
-        action=Merge(),
-        workspace_id=workspace_id,
-        repo_url=f"git@github.com:{_TEST_REPO.slug()}.git",
-        repo=_TEST_REPO,
-        pr_number=_TEST_PR_NUMBER,
-        status=_mergeable_status(),
-        state=state,
-        base_branch=_TEST_DEFAULT_BASE_BRANCH,
-        remote_branch=f"awf/{workspace_id}",
-        remote_push_url=f"git@github.com:{_TEST_REPO.slug()}.git",
-        compose_project="proj",
-        compose_file=tmp_path / "compose.yml",
-        monitor_log=None,
-    )
-
-    assert terminal is False
-    assert gh.merge_calls == ["squash"]
-    # The exhausted-transient counter survives in memory and in the DB instead of
-    # being wiped like a deterministic blocker; the next poll therefore exhausts
-    # immediately rather than re-spending a full bounded budget.
-    counter_key = "__awf_forge_transient_retry_count:merge_pr"
-    assert state.threads_addressed_ids.get(counter_key) == "1"
-    async with factory() as s:
-        ws = await WorkspaceRepository(s).get(workspace_id)
-        assert ws is not None
-        assert (ws.monitor_threads_addressed or {}).get(counter_key) == "1"
-
-
-@pytest.mark.unit
-async def test_unclassified_first_merge_failure_notifies_without_method_mismatch(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """An unclassified merge failure notifies as a regular merge blocker."""
-    gh = _MergeMethodClient(
-        repo_methods=("merge", "squash"),
-        branch_methods=("merge", "squash"),
-        merge_results=[
-            GitHubClientError(
-                operation="gh pr merge",
-                returncode=1,
-                stderr="GraphQL: Pull request could not be merged with this method.",
-            ),
-            "MERGESHA789",
-        ],
-    )
-
-    terminal, state, _sleep, _workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    assert gh.merge_calls == ["squash"]
-    assert len(gh.comments) == 1
-    assert "GitHub rejected the merge attempt" in gh.comments[0]
-    assert "MERGE_METHOD_MISMATCH" not in gh.comments[0]
-    assert not any(
-        key.startswith("__awf_merge_method_blocked__:") for key in state.threads_addressed_ids
-    )
-
-
-@pytest.mark.unit
-async def test_merge_blocker_fallback_sets_attention_flag(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """A deterministic merge blocker (e.g. branch protection) notifies a human
-    directly from the merge loop, so it must also stamp the awaiting-human
-    attention flag (#659). ``decide()`` keeps returning ``Merge`` for this PR, so
-    without a set here the escalation signal would never be surfaced at all."""
-    gh = _MergeMethodClient(
-        repo_methods=("merge", "squash"),
-        branch_methods=("merge", "squash"),
-        merge_results=[
-            GitHubClientError(
-                operation="gh pr merge",
-                returncode=1,
-                stderr="GraphQL: Pull request could not be merged with this method.",
-            ),
-        ],
-    )
-
-    terminal, _state, _sleep, workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    assert len(gh.comments) == 1
-    async with factory() as session:
-        ws = await WorkspaceRepository(session).get(workspace_id)
-        assert ws is not None
-        assert ws.awaiting_human_since is not None
-        assert ws.awaiting_human_reason is not None
-        assert "GitHub rejected the merge attempt" in ws.awaiting_human_reason
-
-
-@pytest.mark.unit
-async def test_merge_blocker_fallback_keeps_attention_since_stable_across_polls(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """A branch-protection merge blocker leaves ``decide()`` returning ``Merge``
-    every poll, so the merge loop re-sets the attention flag each cycle. The
-    top-of-``_execute`` attention clear must NOT wipe the persisted episode start
-    on a ``Merge`` action — otherwise ``awaiting_human_since`` is reset to ``now``
-    on every poll and the operator-facing "awaiting human for N" timer never ages
-    while the operator is still blocked (#659)."""
-    workspace_id = await seed_monitoring_workspace(factory)
-    sleep_fn = RecordedSleep()
-    gh = _MergeMethodClient(
-        repo_methods=("merge", "squash"),
-        branch_methods=("merge", "squash"),
-        merge_results=[
-            GitHubClientError(
-                operation="gh pr merge",
-                returncode=1,
-                stderr="GraphQL: Pull request could not be merged with this method.",
-            ),
-        ],
-    )
-    gh.expect_context(
-        repo=_TEST_REPO,
-        pr_number=_TEST_PR_NUMBER,
-        base_branch=_TEST_DEFAULT_BASE_BRANCH,
-    )
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=FakeAdapter(),
-        sleep_fn=sleep_fn,
-        worktrees_root=tmp_path / "worktrees",
-        gh=gh,
-    )
-
-    # Seed a STABLE episode start from an earlier poll: the operator has already
-    # been blocked since well before this poll runs.
-    episode_start = datetime(2026, 1, 1, 12, 0, 0, tzinfo=UTC)
-    async with factory() as session:
-        await WorkspaceRepository(session).set_workspace_attention(
-            workspace_id, reason="prior escalation", now=episode_start
-        )
-        await session.commit()
-
-    terminal = await runner._execute(
-        action=Merge(),
-        workspace_id=workspace_id,
-        repo_url=f"git@github.com:{_TEST_REPO.slug()}.git",
-        repo=_TEST_REPO,
-        pr_number=_TEST_PR_NUMBER,
-        status=_mergeable_status(),
-        state=MonitorState(),
-        base_branch=_TEST_DEFAULT_BASE_BRANCH,
-        remote_branch=f"awf/{workspace_id}",
-        remote_push_url=f"git@github.com:{_TEST_REPO.slug()}.git",
-        compose_project="proj",
-        compose_file=tmp_path / "compose.yml",
-        monitor_log=None,
-    )
-
-    assert terminal is False
-    async with factory() as session:
-        ws = await WorkspaceRepository(session).get(workspace_id)
-        assert ws is not None
-        # The episode start must be PRESERVED (COALESCE), not reset to this poll's
-        # ``now`` by the top-of-_execute clear running before the merge re-set.
-        assert ws.awaiting_human_since == episode_start
-        # The reason still refreshes to the latest escalation.
-        assert ws.awaiting_human_reason is not None
-        assert "GitHub rejected the merge attempt" in ws.awaiting_human_reason
-
-
-@pytest.mark.unit
-async def test_merge_blocker_fallback_notification_transient_error_retries_then_clears(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """A transient post_comment blip while notifying about a *deterministic* merge
-    blocker must run the bounded forge-retry path (wait + persist the
-    ``post_human_notification`` counter) instead of escaping the merge loop, and a
-    later successful post must clear that budget — mirroring the merge-method
-    preflight notification arm rather than calling ``_post_human_notification_once``
-    naked."""
-    workspace_id = await seed_monitoring_workspace(factory)
-    sleep_fn = RecordedSleep()
-
-    def deterministic_merge_error() -> GitHubClientError:
-        return GitHubClientError(
-            operation="gh pr merge",
-            returncode=1,
-            stderr="GraphQL: Pull request could not be merged with this method.",
-        )
-
-    gh = _MergeMethodClient(
-        repo_methods=("merge", "squash"),
-        branch_methods=("merge", "squash"),
-        merge_results=[deterministic_merge_error()],
-        post_comment_error=GitHubClientError(
-            operation="gh pr comment",
-            returncode=1,
-            stderr="HTTP 502 Bad Gateway",
-        ),
-    )
-    gh.expect_context(
-        repo=_TEST_REPO,
-        pr_number=_TEST_PR_NUMBER,
-        base_branch=_TEST_DEFAULT_BASE_BRANCH,
-    )
-    runner = make_runner(
-        factory=factory,
-        cmd=FakeCommandRunner(),
-        adapter=FakeAdapter(),
-        sleep_fn=sleep_fn,
-        worktrees_root=tmp_path / "worktrees",
-        gh=gh,
-    )
-    state = MonitorState()
-
-    async def _run_merge() -> bool | None:
-        return await runner._execute(
-            action=Merge(),
-            workspace_id=workspace_id,
-            repo_url=f"git@github.com:{_TEST_REPO.slug()}.git",
-            repo=_TEST_REPO,
-            pr_number=_TEST_PR_NUMBER,
-            status=_mergeable_status(),
-            state=state,
-            base_branch=_TEST_DEFAULT_BASE_BRANCH,
-            remote_branch=f"awf/{workspace_id}",
-            remote_push_url=f"git@github.com:{_TEST_REPO.slug()}.git",
-            compose_project="proj",
-            compose_file=tmp_path / "compose.yml",
-            monitor_log=None,
-        )
-
-    counter_key = "__awf_forge_transient_retry_count:post_human_notification"
-
-    # Phase 1: deterministic merge blocker + transient 502 on the notification post.
-    # The wrapped retry path waits and re-polls instead of letting the 502 escape,
-    # and persists the bounded ``post_human_notification`` counter.
-    terminal = await _run_merge()
-    assert terminal is False
-    assert gh.merge_calls == ["squash"]
-    assert gh.comments == []
-    assert sleep_fn.calls == [5]
-    assert state.threads_addressed_ids.get(counter_key) == "1"
-
-    # Phase 2: a later poll whose notification post succeeds clears the budget so a
-    # recovered one-off blip never accumulates toward the bounded retry count.
-    gh.post_comment_error = None
-    gh.merge_results = [deterministic_merge_error()]
-    terminal = await _run_merge()
-    assert terminal is False
-    assert gh.merge_calls == ["squash", "squash"]
-    assert len(gh.comments) == 1
-    assert "GitHub rejected the merge attempt" in gh.comments[0]
-    assert counter_key not in state.threads_addressed_ids
-    assert sleep_fn.calls == [5, 60]
-
-
-@pytest.mark.unit
-async def test_mismatched_first_merge_rejection_notifies_without_method_rotation(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """A rejection naming a different method is not enough to rotate methods."""
-    gh = _MergeMethodClient(
-        repo_methods=("merge", "squash"),
-        branch_methods=("merge", "squash"),
-        merge_results=[
-            GitHubClientError(
-                operation="gh pr merge",
-                returncode=1,
-                stderr="GraphQL: Merge commits are not allowed on this repository.",
-            ),
-            "MERGESHA999",
-        ],
-    )
-
-    terminal, state, _sleep, _workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    assert gh.merge_calls == ["squash"]
-    assert len(gh.comments) == 1
-    assert "GitHub rejected the merge attempt" in gh.comments[0]
-    assert "MERGE_METHOD_MISMATCH" not in gh.comments[0]
-    assert not any(
-        key.startswith("__awf_merge_method_blocked__:") for key in state.threads_addressed_ids
-    )
-
-
-@pytest.mark.unit
-async def test_mismatched_last_merge_rejection_notifies_without_method_blocker(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """A rejection naming a different method does not persist a method blocker."""
-    gh = _MergeMethodClient(
-        repo_methods=("squash",),
-        branch_methods=("squash",),
-        merge_results=[
-            GitHubClientError(
-                operation="gh pr merge",
-                returncode=1,
-                stderr="GraphQL: Merge commits are not allowed on this repository.",
-            )
-        ],
-    )
-
-    terminal, state, sleep_fn, _workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    assert gh.merge_calls == ["squash"]
-    assert len(gh.comments) == 1
-    assert "GitHub rejected the merge attempt" in gh.comments[0]
-    assert "MERGE_METHOD_MISMATCH" not in gh.comments[0]
-    assert sleep_fn.calls == [60]
-    assert not any(
-        key.startswith("__awf_merge_method_blocked__:") for key in state.threads_addressed_ids
-    )
-
-
-@pytest.mark.unit
-async def test_method_rejection_notification_truncates_long_github_detail(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """Final method-blocker notifications preserve fields and cap GitHub detail."""
-    long_detail = " ".join(f"detail{i:03d}" for i in range(80))
-    gh = _MergeMethodClient(
-        repo_methods=("squash",),
-        branch_methods=("squash",),
-        merge_results=[
-            GitHubClientError(
-                operation="gh pr merge",
-                returncode=1,
-                stderr=(
-                    f"GraphQL: Squash merges are not allowed on this repository. {long_detail}"
-                ),
-            )
-        ],
-    )
-
-    terminal, _state, _sleep, _workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    assert len(gh.comments) == 1
-    comment = gh.comments[0]
-    assert "attempted=squash; effective_allowed=squash" in comment
-    assert "detail000" in comment
-    assert "detail079" not in comment
-    github_detail = comment.split("GitHub reported: ", 1)[1].split("\n\n", 1)[0]
-    assert len(github_detail.removesuffix(".")) <= 240
-
-
-@pytest.mark.unit
-async def test_unclassified_single_merge_failure_notifies_without_method_blocker(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """An unclassified single-method failure does not record a method blocker."""
-    gh = _MergeMethodClient(
-        repo_methods=("squash",),
-        branch_methods=("squash",),
-        merge_results=[
-            GitHubClientError(
-                operation="gh pr merge",
-                returncode=1,
-                stderr="GraphQL: Pull request could not be merged with this method.",
-            )
-        ],
-    )
-
-    terminal, state, sleep_fn, _workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    assert gh.merge_calls == ["squash"]
-    assert len(gh.comments) == 1
-    assert "GitHub rejected the merge attempt" in gh.comments[0]
-    assert "MERGE_METHOD_MISMATCH" not in gh.comments[0]
-    assert sleep_fn.calls == [60]
-    assert not any(
-        key.startswith("__awf_merge_method_blocked__:") for key in state.threads_addressed_ids
-    )
-
-
-@pytest.mark.unit
-async def test_deterministic_bitbucket_merge_failure_notifies_and_keeps_polling(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """A permanent Bitbucket merge fault notifies a human instead of terminating.
-
-    Bitbucket workspaces merge through ``BitbucketClient.merge_pr``, which raises
-    ``BitbucketClientError`` (not ``GitHubClientError``). A deterministic failure
-    (branch restrictions, unresolved tasks, a 4xx) must follow the GitHub
-    merge-blocker behaviour — post a human notification and keep polling — rather
-    than escaping ``_attempt_merge_method`` and terminating the workspace.
-    """
-    gh = _MergeMethodClient(
-        repo_methods=("squash",),
-        branch_methods=("squash",),
-        merge_results=[
-            BitbucketClientError(
-                operation="merge_pr",
-                status=403,
-                body="merge checks have not passed",
-            )
-        ],
-    )
-
-    terminal, state, sleep_fn, _workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    assert gh.merge_calls == ["squash"]
-    assert len(gh.comments) == 1
-    assert "Bitbucket rejected the merge attempt" in gh.comments[0]
-    assert sleep_fn.calls == [60]
-    assert not any(
-        key.startswith("__awf_merge_method_blocked__:") for key in state.threads_addressed_ids
-    )
-
-
-@pytest.mark.unit
-async def test_deterministic_bitbucket_merge_failure_forwards_specific_reason_code(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """The failed merge operation forwards ``exc.reason_code`` end-to-end.
-
-    A specific code such as ``BITBUCKET_MERGE_METHOD_UNSUPPORTED`` (raised when a
-    merge method maps to no Bitbucket strategy) must surface in the operation
-    record and audit event rather than being flattened to a generic
-    ``BITBUCKET_MERGE_FAILED`` — otherwise an operator inspecting events has to
-    read the prose ``error_message`` to recover the real cause.
-    """
-    gh = _MergeMethodClient(
-        repo_methods=("squash",),
-        branch_methods=("squash",),
-        merge_results=[
-            BitbucketClientError(
-                operation="merge_pr",
-                status=None,
-                body="unsupported merge method for Bitbucket: 'rebase'",
-                reason_code=BITBUCKET_MERGE_METHOD_UNSUPPORTED,
-            )
-        ],
-    )
-
-    terminal, _state, _sleep_fn, workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    async with factory() as session:
-        operations = await OperationRepository(session).list_for_workspace(
-            workspace_id,
-            operation_type=OperationType.monitor_state,
-        )
-        events = await WorkspaceEventRepository(session).list(workspace_id=workspace_id)
-
-    failed_ops = [op for op in operations if op.status == OperationStatus.failed.value]
-    assert failed_ops, "expected a failed merge operation to be recorded"
-    assert all(op.error_code == BITBUCKET_MERGE_METHOD_UNSUPPORTED for op in failed_ops)
-    assert not any(op.error_code == "BITBUCKET_MERGE_FAILED" for op in failed_ops)
-
-    merge_failed_events = [
-        event
-        for event in events
-        if isinstance(event.payload, dict)
-        and event.payload.get("action") == "merge"
-        and event.payload.get("outcome") == "failed"
-    ]
-    assert merge_failed_events, "expected a failed merge audit event"
-    assert all(
-        event.payload.get("reason_code") == BITBUCKET_MERGE_METHOD_UNSUPPORTED
-        for event in merge_failed_events
-    )
-
-
-@pytest.mark.unit
-async def test_transient_bitbucket_merge_failure_waits_without_notify(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """A transient Bitbucket merge blip waits and re-polls without notifying."""
-    gh = _MergeMethodClient(
-        repo_methods=("squash",),
-        branch_methods=("squash",),
-        merge_results=[
-            BitbucketClientError(
-                operation="merge_pr",
-                status=429,
-                body="rate limited",
-                reason_code=BITBUCKET_RATE_LIMITED,
-            )
-        ],
-    )
-
-    terminal, _state, sleep_fn, _workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    assert gh.merge_calls == ["squash"]
-    assert gh.comments == []
-    assert sleep_fn.calls == [5]
-
-
-@pytest.mark.unit
-async def test_in_progress_bitbucket_merge_does_not_record_failed_operation(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """A 409 ``BITBUCKET_MERGE_IN_PROGRESS`` cancels (never fails) the operation.
-
-    The reason code is transient: ``_wait_after_transient_bitbucket_error`` keeps
-    the monitor polling and a later ``fetch_pr_status`` observes the in-flight
-    merge's MERGED state, completing the workspace successfully. Recording a
-    permanent ``BITBUCKET_MERGE_FAILED`` operation here would leave an inconsistent
-    audit trail (operation "failed" while the workspace completes), so the failure
-    record is omitted for this reason code while still waiting and re-polling.
-
-    The running merge-attempt operation must still reach a terminal state: if the
-    in-flight merge completes before the next loop re-enters ``Merge`` the monitor
-    short-circuits to completion and never finishes this operation, orphaning it as
-    ``running`` forever. The attempt is therefore cancelled before returning the
-    transient blocker.
-    """
-    gh = _MergeMethodClient(
-        repo_methods=("squash",),
-        branch_methods=("squash",),
-        merge_results=[
-            BitbucketClientError(
-                operation="merge_pr",
-                status=409,
-                body="merge already in progress",
-                reason_code=BITBUCKET_MERGE_IN_PROGRESS,
-            )
-        ],
-    )
-
-    terminal, _state, sleep_fn, workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    assert gh.merge_calls == ["squash"]
-    assert gh.comments == []
-    assert sleep_fn.calls == [5]
-
-    async with factory() as session:
-        operations = await OperationRepository(session).list_for_workspace(
-            workspace_id,
-            operation_type=OperationType.monitor_state,
-        )
-        audit_events = await WorkspaceEventRepository(session).list(
-            workspace_id=workspace_id,
-            event_type="workspace.audit.merge_result",
-            limit=20,
-        )
-    assert not any(op.error_code == "BITBUCKET_MERGE_FAILED" for op in operations)
-    assert not any(op.status == OperationStatus.failed.value for op in operations)
-    # The merge-attempt operation must be terminal (cancelled), not orphaned as
-    # ``running`` — otherwise a later ShortCircuitCompleted leaves it stuck.
-    merge_ops = [
-        op
-        for op in operations
-        if isinstance(op.payload, dict) and op.payload.get("reason_code") == "MERGE"
-    ]
-    assert merge_ops, "expected a merge-attempt operation to be recorded"
-    assert all(op.status == OperationStatus.cancelled.value for op in merge_ops)
-    assert not any(op.status == OperationStatus.running.value for op in operations)
-    # The cancellation must leave an audit breadcrumb so operators can tell
-    # "superseded by an in-flight merge" apart from an unexplained cancelled
-    # operation — every other merge arm (GitHub, Bitbucket failure, success)
-    # records a merge_result event, so this transient arm must too.
-    in_progress_events = [
-        event for event in audit_events if event.reason_code == BITBUCKET_MERGE_IN_PROGRESS
-    ]
-    assert len(in_progress_events) == 1
-    audit_payload = in_progress_events[0].payload
-    assert isinstance(audit_payload, dict)
-    assert audit_payload["action"] == "merge"
-    assert audit_payload["outcome"] == "cancelled"
-
-
-async def test_merge_task_timeout_cancels_operation_and_keeps_polling(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    """An exhausted async-merge poll budget cancels (never fails) the operation.
-
-    ``BitbucketClient`` raises ``BITBUCKET_MERGE_TASK_TIMEOUT`` when the bounded
-    poll budget is exhausted while the merge task is still PENDING. The merge may
-    still complete server-side, so this is treated exactly like
-    ``BITBUCKET_MERGE_IN_PROGRESS``: the attempt operation is cancelled (not
-    failed), no "merge rejected" comment is posted, and the monitor waits and
-    re-polls. Misclassifying it as deterministic would post a spurious notification
-    and leave a permanently-failed operation for a merge that succeeds.
-    """
-    gh = _MergeMethodClient(
-        repo_methods=("squash",),
-        branch_methods=("squash",),
-        merge_results=[
-            BitbucketClientError(
-                operation="bitbucket merge_pr (task-status)",
-                status=None,
-                body="Bitbucket merge task did not complete within 30 polls",
-                reason_code=BITBUCKET_MERGE_TASK_TIMEOUT,
-            )
-        ],
-    )
-
-    terminal, _state, sleep_fn, workspace_id = await _execute_merge(
-        factory=factory,
-        tmp_path=tmp_path,
-        gh=gh,
-    )
-
-    assert terminal is False
-    assert gh.merge_calls == ["squash"]
-    # No "merge rejected" notification while the async merge may still be running.
-    assert gh.comments == []
-    assert sleep_fn.calls == [5]
-
-    async with factory() as session:
-        operations = await OperationRepository(session).list_for_workspace(
-            workspace_id,
-            operation_type=OperationType.monitor_state,
-        )
-        audit_events = await WorkspaceEventRepository(session).list(
-            workspace_id=workspace_id,
-            event_type="workspace.audit.merge_result",
-            limit=20,
-        )
-    assert not any(op.status == OperationStatus.failed.value for op in operations)
-    merge_ops = [
-        op
-        for op in operations
-        if isinstance(op.payload, dict) and op.payload.get("reason_code") == "MERGE"
-    ]
-    assert merge_ops, "expected a merge-attempt operation to be recorded"
-    assert all(op.status == OperationStatus.cancelled.value for op in merge_ops)
-    assert not any(op.status == OperationStatus.running.value for op in operations)
-    # The cancellation breadcrumb carries the timeout reason code so operators can
-    # tell it apart from the 409 already-in-flight case.
-    timeout_events = [
-        event for event in audit_events if event.reason_code == BITBUCKET_MERGE_TASK_TIMEOUT
-    ]
-    assert len(timeout_events) == 1
-    audit_payload = timeout_events[0].payload
-    assert isinstance(audit_payload, dict)
-    assert audit_payload["outcome"] == "cancelled"
